requirements.txt
google-genai
streamlit
python-dotenv
pypdf
fpdf
supabase